Making a lot of changes to how data is stored and passed between the services and the activities. Previously, we just used the database interface to [redundantly] retrieve clipped items. This was slow and ineffiient.
I've started moving to inter-process communication with Messenger instances. There is a lot of perceived overhead in this, but I think that in the end it will work out much more appropriately. In addition, the ClippedItem now is Parcelable -- this is very fast for inter-process communication.

From here, we need to start stripping out access to the database in places other than the background service. If the other activities need any information from the database, they must send a message to the background service asking for it. This will eliminate race conditions associated with the service adding / removing and the interface requesting.

Note that the application is still using the database handler, but the messaging interface is set up. We just have to utilize it. This commit works.